backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
psycopg2-binary==2.9.9
pgvector==0.2.4
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
numpy==1.24.3
python-jose==3.3.0
supabase==2.8.1